DAO com inserção do produto
