+++ loja-virtual-repository/src/br/com/alura/jdbc/dao/ProdutoDAO.java
@@ -1,0 +1,40 @@
+package br.com.alura.jdbc.dao;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+
+import br.com.alura.jdbc.modelo.Produto;
+
+/*
+ * Data Access Object ou DAO é quem trabalha com as informações do nosso objeto no 
+ * banco de dados.
+ */
+public class ProdutoDAO {
+
+	private Connection connection;
+	
+	public ProdutoDAO(Connection connection) {
+		this.connection = connection;
+	}
+	
+	public void salvar(Produto produto) throws SQLException {
+		String sql = "INSERT INTO PRODUTO (NOME, DESCRICAO) VALUES (?, ?)";
+		
+		try (PreparedStatement pstm = connection.prepareStatement(
+				sql, Statement.RETURN_GENERATED_KEYS)) {
+			pstm.setString(1, produto.getNome());
+			pstm.setString(2, produto.getDescricao());
+			
+			pstm.execute();
+			
+			try (ResultSet rst = pstm.getGeneratedKeys()){
+				while (rst.next()) {
+					produto.setId(rst.getInt(1));
+				}
+			}
+		}
+	}
+}
